fix: fix typer på nyhetssiden
