--- portal/src/app/(frontend)/nyheter/article.types.ts
@@ -1,45 +1,0 @@
-import type { DocumentType } from "@/sanity/schemas/documents";
-
-type ArticleType = {
-    value: string;
-    sanityType: DocumentType;
-    label: string;
-};
-
-export const ARTICLE_TYPES: readonly ArticleType[] = [
-    {
-        value: "blogg",
-        sanityType: "jokul_blog_post",
-        label: "Blogg",
-    },
-    {
-        value: "komponent",
-        sanityType: "jokul_component",
-        label: "Komponenter",
-    },
-    {
-        value: "fundament",
-        sanityType: "jokul_fundamentals",
-        label: "Fundamenter",
-    },
-    {
-        value: "release-notes",
-        sanityType: "jokul_release_notes",
-        label: "Release notes",
-    },
-    {
-        value: "monster",
-        sanityType: "jokul_monster",
-        label: "Mønster",
-    },
-] as const;
-
-const articleTypeValues = new Set<string>(
-    ARTICLE_TYPES.map((type) => type.value),
-);
-
-export function isArticleTypeValue(
-    value: string,
-): value is ArticleType["value"] {
-    return articleTypeValues.has(value);
-}
